+++ ai/turnplayer/utils.go
@@ -20,7 +20,7 @@ func GenBestStaticTurn(g *game.Game, p AITurnPlayer, playerIdx int) *move.Move {
 	// Add an exchange only if there are 7 or more tiles in the bag.
 	// in case we don't have full rack info:
 	oppRack := g.RackFor(1 - playerIdx)
-	unseen := oppRack.NumTiles() + uint8(g.Bag().TilesRemaining())
+	unseen := int(oppRack.NumTiles()) + g.Bag().TilesRemaining()
 	exchAllowed := unseen-game.RackTileLimit >= game.ExchangeLimit
 	mg.GenAll(g.RackFor(playerIdx), exchAllowed)
 	return mg.Plays()[0]
